requirements.txt
# Core dependencies
torch>=2.0.0
torchaudio
transformers>=4.30.0
scipy
diffusers>=0.20.0
accelerate
requests
Pillow>=9.0.0

# Web framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# Audio processing
librosa
soundfile

# OpenAI API
openai==0.28.1
python-dotenv

# Optional: For advanced features
# datasets  # Uncomment if using fine-tuning
# wandb     # Uncomment if using training monitoring